+++ Composer/packages/lib/shared/src/index.ts
@@ -11,3 +11,4 @@ export * from './dialogFactory';
 export * from './promptTabs';
 export * from './appschema';
 export * from './types';
+export * from './lgUtils';
